LRQA-20643 Merge WebDriverToSeleniumBridge into BaseWebDriverImpl
